agent_worker/requirements.txt
# LiveKit Agents framework
livekit-agents~=1.4
livekit-plugins-silero~=1.4
livekit-plugins-langchain~=1.4

# STT — FasterWhisper (CTranslate2-based, ~4x faster than openai-whisper)
faster-whisper>=1.0.0

# TTS — Kokoro
kokoro>=0.3.0
soundfile>=0.12.0

# ML runtime
torch>=2.0.0
numpy<2.0

# LangGraph Agent
langgraph==0.2.0
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0

# Config
PyYAML==6.0.2
